refactor(ops): slim big5 controller with action/query services
